Add test harness
Adds a test harness in the form of docker-compose. An enviornment is
brought up then tests are run against it before it is brought down.

For now only a very simple test has been added. (Checking that Gremlin
agrees that 1+1 = 2.) In the future we should add more advanced tests
that test that both backends are working.